Set fsGroup instead of trying to chmod in init container
The init container needs to run as root, which is not always possible,
so having the storage class provide group write access should be more
general.

This patch tries to grab the GID from runAsGroup. If that isn't set, it
tries runAsUser because the vcap user typically uses the same UID as GID.

RunAsUser is currently being set by Eirini; if it weren't we fall back to
a hardcoded '2000' as the best remaining guess available...